test/Add Java Web Signature test
